% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{criar_serie_municipio}
\alias{criar_serie_municipio}
\title{Série temporal de vítimas}
\usage{
criar_serie_municipio(tab, munip)
}
\arguments{
\item{tab}{Base de dados do Sinesp por município.}

\item{munip}{Uma string com o nome do município assim como na base.}
}
\description{
Cria uma série temporal de vítimas de um município. Fonte: Sinesp.
}
