% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{gerar_grafico_serie}
\alias{gerar_grafico_serie}
\title{Gera um grafico de serie}
\usage{
gerar_grafico_serie(tab, municipio_nome)
}
\arguments{
\item{tab}{A base sinesp por municipio}

\item{municipio_nome}{Uma string com nome de um
municipio da base SINESP}
}
\value{
um grafico ggplot
}
\description{
Gera um grafico de serie para um municipio da base SINESP
}
