% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graficos.R
\name{criar_serie_municipio}
\alias{criar_serie_municipio}
\title{Gera um grafico de serie}
\usage{
criar_serie_municipio(tab, munip)
}
\arguments{
\item{tab}{uma tabela}

\item{nome_muinicipio}{nome de um nunuciipo da tabela}
}
\value{
um grafico
}
\description{
Gera um grafico de serie
}
